add parameter to EthereumApplication which allows to reset the datadir

+++ android-geth/src/main/java/com/sqli/blockchain/android_geth/EthereumService.java
@@ -22,6 +22,7 @@ public class EthereumService extends Service {
     public static final String GETH_IPC_FILE = "geth.ipc";
     public static final String GETH_GENESIS_FILE = "genesis.json";
     public static final String GETH_BOOTNODE_FILE = "static-nodes.json";
+    public static final String RESET_DATADIR_FLAG = "RESET_DATADIR_FLAG";
 
     public static String dataDir;
     private final IBinder mBinder = new LocalBinder();
@@ -38,11 +39,15 @@ public class EthereumService extends Service {
         dataDir = getFilesDir().getAbsolutePath();
         callbacks = new ArrayList<>();
 
-        deleteDatadir();
+        boolean resetDatadir = intent.getBooleanExtra(RESET_DATADIR_FLAG,false);
+
 
         try {
 
-            Utils.saveAssetOnStorage(getBaseContext(), GETH_GENESIS_FILE, dataDir);
+            if( resetDatadir ) {
+                deleteDatadir();
+                Utils.saveAssetOnStorage(getBaseContext(), GETH_GENESIS_FILE, dataDir);
+            }
 
             final StringBuilder gethParams = new StringBuilder();
             gethParams.append("--fast").append(" ");
